+++ src/ui/menu.cpp
@@ -112,7 +112,7 @@ void MenuList::UpdateAnime(const anime::Item* anime_item) {
       item.checked = false;
       item.def = false;
     }
-    int item_index = anime_item->GetMyStatus();
+    int item_index = static_cast<int>(anime_item->GetMyStatus());
     if (item_index - 1 < static_cast<int>(menu->items.size())) {
       menu->items[item_index - 1].checked = true;
       menu->items[item_index - 1].def = true;
